Changed datatype of Vector2D to be double

+++ src/Engine/Utility/vector2D.cpp
@@ -11,7 +11,7 @@
 
 Vector2D::Vector2D() : x(0), y(0) {}
 
-Vector2D::Vector2D(float x, float y) : x(x), y(y) {}
+Vector2D::Vector2D(double x, double y) : x(x), y(y) {}
 
 Vector2D operator+(const Vector2D &v1, const Vector2D &v2) {
   return Vector2D(v1.x + v2.x, v1.y + v2.y);
@@ -47,22 +47,22 @@ Vector2D &Vector2D::operator/=(const Vector2D &other) {
   return *this;
 }
 
-Vector2D &Vector2D::operator+=(const float scaler) {
+Vector2D &Vector2D::operator+=(const double scaler) {
   this->x += scaler;
   this->y += scaler;
   return *this;
 }
-Vector2D &Vector2D::operator-=(const float scaler) {
+Vector2D &Vector2D::operator-=(const double scaler) {
   this->x -= scaler;
   this->y -= scaler;
   return *this;
 }
-Vector2D &Vector2D::operator*=(const float scaler) {
+Vector2D &Vector2D::operator*=(const double scaler) {
   this->x *= scaler;
   this->y *= scaler;
   return *this;
 }
-Vector2D &Vector2D::operator/=(const float scaler) {
+Vector2D &Vector2D::operator/=(const double scaler) {
   this->x /= scaler;
   this->y /= scaler;
   return *this;
